add upgrade to dimensions-config pfs version=2
This new version is required for LSST v28. It involves changing
the dither dimension to an integer from a float, and setting the
length on various strings. I'm not bothering to code the
downgrade, since I don't think we'll ever use it; it should be
straightforward to write, based on the upgrade, if it ever
becomes necessary.